Program works for small numbers, but...
there seems to be an issue with larger numbers, and so far it's probably NOT  a hashmap issues, as there are too few frames for the no of buckets
I have. I'll have to investigate further to see if a function (NOT a loop) has a problem, because loops seem(?) to be working fine.